+++ src/dse_polars/__init__.py
@@ -8,10 +8,11 @@ except PackageNotFoundError:
 
 from .dse import DSE
 from .texts import DSEPassages, ctsurn_contains
-from .images import CitableIIIFService, urn2info_url, info_url2urn, roi, strip_roi
+from .images import CitableIIIFService, urn2info_url, info_url2urn, roi, strip_roi, ptinrect, rois, urn2image_url
 from .urnutils import passagecomponent_re
 
 __all__ = ["DSE", 
+           "passagecomponent_re",
            "DSEPassages",
            "ctsurn_contains",
          
@@ -20,4 +21,7 @@ __all__ = ["DSE",
            "urn2info_url",
            "info_url2urn",
            "roi","strip_roi",
-           "passagecomponent_re"]
+            "ptinrect",
+            "rois",
+            "urn2image_url"
+        ]